+++ Instagram Clone Coding/Instagram Clone Coding/Sources/Model/ImageModel.swift
@@ -1,0 +1,8 @@
+//
+//  ImageModel.swift
+//  Instagram Clone Coding
+//
+//  Created by 김지민 on 2022/05/15.
+//
+
+import Foundation
